item option and modifier data oopsy #3

+++ src/models/catalog_models.rs
@@ -77,8 +77,8 @@ pub struct CatalogItemTextModifier {
 pub struct CatalogItemModifierList {
     pub modifier_list_id: String,
     pub modifier_value_ids: Vec<String>,
-    pub min_selected_modifiers: i32,
-    pub max_selected_modifiers: i32
+    pub min_selected_modifiers: i64,
+    pub max_selected_modifiers: i64
 }
 
 #[derive(Serialize, Deserialize, PartialEq, Hash)]
